requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
convertdate==2.4.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.6
django-autocomplete-light==3.9.4
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-filter==22.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-mptt==0.13.4
django-phonenumber-field==6.3.0
django-resized==1.0.1
django-tables2==2.4.1
django-tables2-column-shifter==2.0.3
docopt==0.6.2
et-xmlfile==1.1.0
flower==1.1.0
hijri-converter==2.2.4
holidays==0.14.2
humanize==4.2.3
kombu==5.2.4
korean-lunar-calendar==0.2.1
loguru==0.6.0
MarkupPy==1.14
num2words==0.5.11
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
phonenumbers==8.12.51
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
PyMeeus==0.5.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.4
pytz==2022.1
PyYAML==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tornado==6.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
